requirements.txt
selenium
requests
beautifulsoup4
Pillow
colorama
